Improvements to game difficulty.

+++ src/teleport.c
@@ -25,7 +25,7 @@ boolean
 goodpos(x, y, mtmp, gpflags)
 int x, y;
 struct monst *mtmp;
-unsigned gpflags;
+unsigned long gpflags;
 {
     struct permonst *mdat = (struct permonst *) 0;
     boolean ignorewater = ((gpflags & MM_IGNOREWATER) != 0);
@@ -116,7 +116,7 @@ enexto_core(cc, xx, yy, mdat, entflags)
 coord *cc;
 register xchar xx, yy;
 struct permonst *mdat;
-unsigned entflags;
+unsigned long entflags;
 {
 #define MAX_GOOD 15
     coord good[MAX_GOOD], *good_ptr;
